Added testing files framework
Still need to setup the browser support for phantomjs to support development on c9.io But for now I will continue work on the UI portion of the project until I can fully support cloud development and testing. Next steps will be to do some simple tests, and then hopefully utilize travis CI for this project.