Added filter list implimentation for stats.ctrl
Now there is a filter add, remove buttons that add and remove filters
which will be applied when searching the database in the stat view.
Next steps will be to work on the other canvas types.
I may move the tab list format below the filters, this way the graph
will change but the filters will stay.