Fixed failing karma tests
Due to the large amount of code/tests removed during the change of
requirements for the stat page, the number of tests dropped. There as
alos changes within the asterank service that required changes in the
asterank spec test classes. I will expand upon the asterank test classes
later, need to work on the actual code first.