Fixed and refactored table view
The table view now is ready to be setup where you can display either the
by day or by size. This will be mainly useful to display the data
differently which is a key part of the application.

I still need to define how the getMonth will work, or how it will
change. I believe a basic get dangerous, and sort by size would be good.
So a double filter would be nice.

I can futher expand on this functionality for all the other views as
well.