Bumped to v1.0.5
Main fixes include UI fixes for critical bugs where the window will
resize to essentially nothing in the X width. This was due to the
dynamic resizing of chrome extensions, page overflow and the mdMenu.

Main fix was to force the page width and replace the builtIn mdMenu with
a basic custom version.